Add class Disk
(#669) mv disk_manager to block_manager
