+++ src/chunkserver/block_manager.cc
@@ -1,49 +1,44 @@
-// Copyright (c) 2015, Baidu.com, Inc. All Rights Reserved.
+// Copyright (c) 2016, Baidu.com, Inc. All Rights Reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
-// Author: yanshiguang02@baidu.com
 
 #include "chunkserver/block_manager.h"
 
-#include <errno.h>
-#include <string.h>
 #include <sys/stat.h>
 #include <sys/vfs.h>
+#include <climits>
 #include <functional>
-#include <algorithm>
 
 #include <gflags/gflags.h>
 #include <leveldb/db.h>
-#include <leveldb/cache.h>
+#include <leveldb/iterator.h>
 #include <common/counter.h>
 #include <common/logging.h>
 #include <common/string_util.h>
 
+#include "chunkserver/disk.h"
 #include "chunkserver/data_block.h"
 #include "chunkserver/file_cache.h"
 
 DECLARE_int32(chunkserver_file_cache_size);
 DECLARE_int32(chunkserver_use_root_partition);
-DECLARE_int32(chunkserver_io_thread_num);
 
 namespace baidu {
 namespace bfs {
 
 extern common::Counter g_blocks;
-extern common::Counter g_data_size;
 extern common::Counter g_find_ops;
 
 BlockManager::BlockManager(const std::string& store_path)
-   : metadb_(NULL),
-     namespace_version_(0), disk_quota_(0) {
-     CheckStorePath(store_path);
-     thread_pool_ = new ThreadPool(FLAGS_chunkserver_io_thread_num);
-     file_cache_ = new FileCache(FLAGS_chunkserver_file_cache_size);
+    : disk_quota_(0) {
+    CheckStorePath(store_path);
+    file_cache_ = new FileCache(FLAGS_chunkserver_file_cache_size);
 }
+
 BlockManager::~BlockManager() {
     MutexLock lock(&mu_);
-    for (BlockMap::iterator it = block_map_.begin(); it != block_map_.end(); ++it) {
+    for (auto it = block_map_.begin(); it != block_map_.end(); ++it) {
         Block* block = it->second;
         if (!block->IsRecover()) {
             CloseBlock(block);
@@ -52,21 +47,18 @@ BlockManager::~BlockManager() {
         }
         block->DecRef();
     }
-    thread_pool_->Stop(true);
-    delete thread_pool_;
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        delete disks_[i];
+    }
     block_map_.clear();
-    delete metadb_;
-    metadb_ = NULL;
     delete file_cache_;
     file_cache_ = NULL;
-}
-int64_t BlockManager::DiskQuota() const{
-    return disk_quota_;
 }
 
 void BlockManager::CheckStorePath(const std::string& store_path) {
     int64_t disk_quota = 0;
-    common::SplitString(store_path, ",", &store_path_list_);
+    std::vector<std::string> store_path_list;
+    common::SplitString(store_path, ",", &store_path_list);
     std::map<std::string, std::string> fs_map;
     std::string fsid_str;
     struct statfs fs_info;
@@ -78,8 +70,8 @@ void BlockManager::CheckStorePath(const std::string& store_path) {
         fs_map[fsid_str] = "Root";
         LOG(INFO, "Root fsid: %s", common::DebugString(fsid_str).c_str());
     }
-    for (uint32_t i = 0; i < store_path_list_.size(); ++i) {
-        std::string& disk_path = store_path_list_[i];
+    for (size_t i = 0; i < store_path_list.size(); ++i) {
+        std::string& disk_path = store_path_list[i];
         disk_path = common::TrimString(disk_path, " ");
         if (disk_path.empty() || disk_path[disk_path.size() - 1] != '/') {
             disk_path += "/";
@@ -97,6 +89,8 @@ void BlockManager::CheckStorePath(const std::string& store_path) {
                 common::HumanReadableString(super_quota).c_str(),
                 common::HumanReadableString(user_quota).c_str());
             disk_quota += user_quota;
+            Disk* disk = new Disk(store_path_list[i], file_cache_, user_quota);
+            disks_.push_back(disk);
         } else {
             if (stat_ret != 0) {
                 LOG(WARNING, "Stat store_path %s fail: %s, ignore it", disk_path.c_str(), strerror(errno));
@@ -104,189 +98,102 @@ void BlockManager::CheckStorePath(const std::string& store_path) {
                 LOG(WARNING, "%s's fsid is same to %s, ignore it",
                     disk_path.c_str(), fs_map[fsid_str].c_str());
             }
-            store_path_list_[i] = store_path_list_[store_path_list_.size() - 1];
-            store_path_list_.resize(store_path_list_.size() - 1);
+            store_path_list[i] = store_path_list[store_path_list.size() - 1];
+            store_path_list.resize(store_path_list.size() - 1);
             --i;
         }
     }
-    std::sort(store_path_list_.begin(), store_path_list_.end());
-    std::vector<std::string>::iterator it
-       = std::unique(store_path_list_.begin(), store_path_list_.end());
-    store_path_list_.resize(std::distance(store_path_list_.begin(), it));
-    LOG(INFO, "%lu store path used.", store_path_list_.size());
-    assert(store_path_list_.size() > 0);
     disk_quota_ = disk_quota;
 }
-const std::string& BlockManager::GetStorePath(int64_t block_id) {
-    return store_path_list_[block_id % store_path_list_.size()];
-}
-/// Load meta from disk
-bool BlockManager::LoadStorage() {
-    MutexLock lock(&mu_);
-    int64_t start_load_time = common::timer::get_micros();
-    leveldb::Options options;
-    options.create_if_missing = true;
-    leveldb::Status s = leveldb::DB::Open(options, store_path_list_[0] + "meta/", &metadb_);
-    if (!s.ok()) {
-        LOG(WARNING, "Load blocks fail: %s", s.ToString().c_str());
-        return false;
-    }
 
-    std::string version_key(8, '\0');
-    version_key.append("version");
-    std::string version_str;
-    s = metadb_->Get(leveldb::ReadOptions(), version_key, &version_str);
-    if (s.ok() && version_str.size() == 8) {
-        namespace_version_ = *(reinterpret_cast<int64_t*>(&version_str[0]));
-        LOG(INFO, "Load namespace %ld", namespace_version_);
+int64_t BlockManager::DiskQuota() const {
+    return disk_quota_;
+}
+
+// TODO: concurrent load
+bool BlockManager::LoadStorage() {
+    bool ret = true;
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        ret = ret && disks_[i]->LoadStorage(std::bind(&BlockManager::AddBlock,
+                                                      this, std::placeholders::_1,
+                                                      std::placeholders::_2));
     }
-    int block_num = 0;
-    leveldb::Iterator* it = metadb_->NewIterator(leveldb::ReadOptions());
-    for (it->Seek(version_key+'\0'); it->Valid(); it->Next()) {
-        int64_t block_id = 0;
-        if (1 != sscanf(it->key().data(), "%ld", &block_id)) {
-            LOG(WARNING, "Unknown key: %s\n", it->key().ToString().c_str());
-            delete it;
+    return ret;
+}
+
+int64_t BlockManager::NameSpaceVersion() const {
+    int64_t version = -1;
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        int64_t tmp = disks_[i]->NameSpaceVersion();
+        if (version == -1) {
+            version = tmp;
+        }
+        if (tmp != version) {
+            return -1;
+        }
+    }
+    return version;
+}
+
+bool BlockManager::SetNameSpaceVersion(int64_t version) {
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        if (!disks_[i]->SetNameSpaceVersion(version)) {
             return false;
         }
-        BlockMeta meta;
-        if (!meta.ParseFromArray(it->value().data(), it->value().size())
-            || meta.block_id() != block_id) {
-            struct OldBlockMeta {
-                int64_t block_id;
-                int64_t block_size;
-                int64_t checksum;
-                int64_t version;
-            };
-            assert(it->value().size() == sizeof(struct OldBlockMeta));
-            OldBlockMeta oldmeta;
-            memcpy(&oldmeta, it->value().data(), it->value().size());
-            assert(oldmeta.block_id == block_id);
-            meta.set_block_id(oldmeta.block_id);
-            meta.set_block_size(oldmeta.block_size);
-            meta.set_checksum(oldmeta.checksum);
-            meta.set_version(oldmeta.version);
-            meta.set_store_path(GetStorePath(block_id));
-            std::string meta_buf;
-            meta.SerializeToString(&meta_buf);
-            metadb_->Put(leveldb::WriteOptions(), it->key(), meta_buf);
-            LOG(INFO, "Old meta info of #%ld has been trans to new meta info", oldmeta.block_id);
-        }
-        std::string file_path = meta.store_path() + Block::BuildFilePath(block_id);
-        if (meta.version() < 0) {
-            LOG(INFO, "Incomplete block #%ld V%ld %ld, drop it",
-                block_id, meta.version(), meta.block_size());
-            metadb_->Delete(leveldb::WriteOptions(), it->key());
-            remove(file_path.c_str());
-            continue;
-        } else {
-            if (std::find(store_path_list_.begin(), store_path_list_.end(), meta.store_path())
-                    == store_path_list_.end()) {
-                LOG(WARNING, "Block #%ld store path %s not in current store configuration, ignore it",
-                    block_id, file_path.c_str());
-                continue;
-            }
-            struct stat st;
-            if (stat(file_path.c_str(), &st) ||
-                st.st_size != meta.block_size() ||
-                access(file_path.c_str(), R_OK)) {
-                LOG(WARNING, "Corrupted block #%ld V%ld size %ld path %s can't access: %s'",
-                    block_id, meta.version(), meta.block_size(), file_path.c_str(),
-                    strerror(errno));
-                metadb_->Delete(leveldb::WriteOptions(), it->key());
-                remove(file_path.c_str());
-                continue;
-            } else {
-                LOG(DEBUG, "Load #%ld V%ld size %ld path %s",
-                    block_id, meta.version(), meta.block_size(), file_path.c_str());
-            }
-        }
-        Block* block = new Block(meta, thread_pool_, file_cache_);
-        block->AddRef();
-        block_map_[block_id] = block;
-        block_num ++;
     }
-    delete it;
-    int64_t end_load_time = common::timer::get_micros();
-    LOG(INFO, "Load %ld blocks, use %ld ms, namespace version: %ld",
-        block_num, (end_load_time - start_load_time) / 1000, namespace_version_);
-    if (namespace_version_ == 0 && block_num > 0) {
-        LOG(WARNING, "Namespace version lost!");
-    }
-    disk_quota_ += g_data_size.Get();
-    return true;
-}
-int64_t BlockManager::NameSpaceVersion() const {
-    return namespace_version_;
-}
-bool BlockManager::SetNameSpaceVersion(int64_t version) {
-    MutexLock lock(&mu_);
-    std::string version_key(8, '\0');
-    version_key.append("version");
-    std::string version_str(8, '\0');
-    *(reinterpret_cast<int64_t*>(&version_str[0])) = version;
-    leveldb::Status s = metadb_->Put(leveldb::WriteOptions(), version_key, version_str);
-    if (!s.ok()) {
-        LOG(WARNING, "SetNameSpaceVersion fail: %s", s.ToString().c_str());
-        return false;
-    }
-    namespace_version_ = version;
-    LOG(INFO, "Set namespace version: %ld", namespace_version_);
     return true;
 }
 
-int64_t BlockManager::ListBlocks(std::vector<BlockMeta>* blocks, int64_t offset, int32_t num) {
-    leveldb::Iterator* it = metadb_->NewIterator(leveldb::ReadOptions());
+// TODO: need test
+int64_t BlockManager::ListBlocks(std::vector<BlockMeta>* blocks, int64_t offset, uint32_t num) {
+    std::vector<leveldb::Iterator*> iters;
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        disks_[i]->Seek(offset, &iters);
+    }
+    if (iters.size() == 0) {
+        return 0;
+    }
+    int32_t idx = -1;
     int64_t largest_id = 0;
-    for (it->Seek(BlockId2Str(offset)); it->Valid(); it->Next()) {
-        int64_t block_id = 0;
-        if (1 != sscanf(it->key().data(), "%ld", &block_id)) {
-            LOG(WARNING, "[ListBlocks] Unknown meta key: %s\n",
-                it->key().ToString().c_str());
-            break;
+    while (blocks->size() < num && iters.size() != 0) {
+        int64_t block_id = FindSmallest(iters, &idx);
+        if (block_id == -1) {
+            return largest_id;
         }
+        auto it = iters[idx];
         BlockMeta meta;
         bool ret = meta.ParseFromArray(it->value().data(), it->value().size());
         assert(ret);
-        //skip blocks not in current configuration
-        if (find(store_path_list_.begin(), store_path_list_.end(), meta.store_path())
-                  == store_path_list_.end()) {
-            continue;
-        }
-        assert(meta.block_id() == block_id);
+        assert(block_id == meta.block_id());
         blocks->push_back(meta);
         largest_id = block_id;
-        // LOG(DEBUG, "List block %ld", block_id);
-        if (--num <= 0) {
-            break;
-        }
+        iters[idx]->Next();
     }
-    delete it;
     return largest_id;
 }
 
-Block* BlockManager::CreateBlock(int64_t block_id, int64_t* sync_time, StatusCode* status) {
+Block* BlockManager::CreateBlock(int64_t block_id, StatusCode* status) {
     BlockMeta meta;
     meta.set_block_id(block_id);
-    meta.set_store_path(GetStorePath(block_id));
-    Block* block = new Block(meta, thread_pool_, file_cache_);
+    Disk* disk = PickDisk(block_id);
+    meta.set_store_path(disk->Path());
+    Block* block = new Block(meta, NULL, file_cache_);
+    // for block_map_
     MutexLock lock(&mu_, "BlockManger::AddBlock", 1000);
-    BlockMap::iterator it = block_map_.find(block_id);
-    if (it != block_map_.end()) {
+    auto ret = block_map_.insert(std::make_pair(block_id, block));
+    if (ret.second) {
+        block->AddRef();
+    } else {
         delete block;
-        if (it->second->IsFinished()) {
+        if (ret.first->second->IsFinished()) {
             *status = kReadOnly;
             return NULL;
         }
         *status = kBlockExist;
-        return it->second;
+        return ret.first->second;
     }
-    // for block_map
-    block->AddRef();
-    block_map_[block_id] = block;
-    // Unlock for write meta & sync
     mu_.Unlock();
-    if (!SyncBlockMeta(meta, sync_time)) {
+    if (!disk->SyncBlockMeta(meta)) {
         delete block;
         *status = kSyncMetaFailed;
         block = NULL;
@@ -299,13 +206,59 @@ Block* BlockManager::CreateBlock(int64_t block_id, int64_t* sync_time, StatusCod
         block->AddRef();
     }
     *status = kOK;
+    LOG(INFO, "CreateBlock %ld on %s", block_id, disk->Path().c_str());
     return block;
+}
+
+bool BlockManager::CloseBlock(Block* block) {
+    return block->Close();
+}
+
+StatusCode BlockManager::RemoveBlock(int64_t block_id) {
+    Block* block = FindBlock(block_id);
+    if (!block) {
+        LOG(INFO, "Try to remove block that does not exist: #%ld ", block_id);
+        return kNsNotFound;
+    } else {
+        MutexLock lock(&mu_, "BlockManager::RemoveBlock erase", 1000);
+        block_map_.erase(block_id);
+        block->DecRef();
+        LOG(INFO, "Remove #%ld meta info done, ref= %ld", block_id, block->GetRef());
+    }
+    StatusCode s = block->SetDeleted();
+    if (s != kOK) {
+        LOG(INFO, "Block #%ld deleted failed, %s", block_id, StatusCode_Name(s).c_str());
+        block->DecRef();
+        return s;
+    }
+    // disk file will be removed in block's deconstrucor
+    file_cache_->EraseFileCache(block->GetFilePath());
+    block->DecRef();
+    return kOK;
+}
+
+// TODO: concurrent & async cleanup
+bool BlockManager::CleanUp(int64_t namespace_version) {
+    for (size_t i = 0; i < disks_.size(); ++i) {
+        if (disks_[i]->NameSpaceVersion() != namespace_version) {
+            if (!disks_[i]->CleanUp()) {
+                return false;
+            }
+        }
+    }
+    LOG(INFO, "CleanUp done");
+    return true;
+}
+
+bool BlockManager::AddBlock(int64_t block_id, Block* block) {
+    MutexLock lock(&mu_);
+    return block_map_.insert(std::make_pair(block_id, block)).second;
 }
 
 Block* BlockManager::FindBlock(int64_t block_id) {
     g_find_ops.Inc();
     MutexLock lock(&mu_, "BlockManger::Find", 1000);
-    BlockMap::iterator it = block_map_.find(block_id);
+    auto it = block_map_.find(block_id);
     if (it == block_map_.end()) {
         // not found
         return NULL;
@@ -315,108 +268,36 @@ Block* BlockManager::FindBlock(int64_t block_id) {
     block->AddRef();
     return block;
 }
-std::string BlockManager::BlockId2Str(int64_t block_id) {
-    char idstr[64];
-    snprintf(idstr, sizeof(idstr), "%13ld", block_id);
-    return std::string(idstr);
-}
-bool BlockManager::SyncBlockMeta(const BlockMeta& meta, int64_t* sync_time) {
-    std::string idstr = BlockId2Str(meta.block_id());
-    leveldb::WriteOptions options;
-    // options.sync = true;
-    int64_t time_start = common::timer::get_micros();
-    std::string meta_buf;
-    meta.SerializeToString(&meta_buf);
-    leveldb::Status s = metadb_->Put(options, idstr, meta_buf);
-    int64_t time_use = common::timer::get_micros() - time_start;
-    if (sync_time) *sync_time = time_use;
-    if (!s.ok()) {
-        Log(WARNING, "Write to meta fail:%s", idstr.c_str());
-        return false;
-    }
-    return true;
-}
-bool BlockManager::CloseBlock(Block* block) {
-    if (!block->Close()) {
-        return false;
-    }
 
-    // Update meta
-    BlockMeta meta = block->GetMeta();
-    return SyncBlockMeta(meta, NULL);
+Disk* BlockManager::PickDisk(int64_t block_id) {
+    return disks_[block_id % disks_.size()];
 }
 
-bool BlockManager::RemoveBlockMeta(int64_t block_id) {
-    char idstr[14];
-    snprintf(idstr, sizeof(idstr), "%13ld", block_id);
-    leveldb::Status s = metadb_->Delete(leveldb::WriteOptions(), idstr);
-    if (!s.ok()) {
-        LOG(WARNING, "Remove #%ld meta info fails: %s", block_id, s.ToString().c_str());
-        return false;
-    }
-    return true;
-}
-StatusCode BlockManager::RemoveBlock(int64_t block_id) {
-    bool meta_removed = RemoveBlockMeta(block_id);
-    Block* block = FindBlock(block_id);
-    if (block == NULL) {
-        LOG(INFO, "Try to remove block that does not exist: #%ld ", block_id);
-        return kNsNotFound;
-    }
-    if (!block->SetDeleted()) {
-        LOG(INFO, "Block #%ld deleted by other thread", block_id);
-        block->DecRef();
-        return kSyncMetaFailed;
-    }
-
-    // disk file will be removed in block's deconstrucor
-    file_cache_->EraseFileCache(block->GetFilePath());
-
-    StatusCode ret = kNsNotFound;
-    if (meta_removed) {
-        MutexLock lock(&mu_, "BlockManager::RemoveBlock erase", 1000);
-        block_map_.erase(block_id);
-        block->DecRef();
-        LOG(INFO, "Remove #%ld meta info done, ref= %ld", block_id, block->GetRef());
-        ret = kOK;
-    }
-    block->DecRef();
-    return ret;
-}
-
-bool BlockManager::RemoveAllBlocks() {
-    LOG(INFO, "RemoveAllBlocks...");
-    if (!RemoveAllBlocksAsync()) {
-        return false;
-    }
-    int count = 0;
-    while (g_blocks.Get() > 0) {
-        if (count++ %1000 == 0) {
-            LOG(INFO, "RemoveAllBlocks wait done now block num: %ld", g_blocks.Get());
-        }
-        usleep(10000);
-    }
-    LOG(INFO, "RemoveAllBlocks done");
-    return true;
-}
-
-bool BlockManager::RemoveAllBlocksAsync() {
-    leveldb::Iterator* it = metadb_->NewIterator(leveldb::ReadOptions());
-    for (it->Seek(BlockId2Str(0)); it->Valid(); it->Next()) {
-        int64_t block_id = 0;
-        if (1 != sscanf(it->key().data(), "%ld", &block_id)) {
-            LOG(FATAL, "[ListBlocks] Unknown meta key: %s\n",
-                it->key().ToString().c_str());
+int64_t BlockManager::FindSmallest(std::vector<leveldb::Iterator*>& iters, int32_t* idx) {
+    int64_t id = LLONG_MAX;
+    for (size_t i = 0; i < iters.size(); ++i) {
+        auto it = iters[i];
+        if (!it->Valid()) {
             delete it;
-            return false;
+            iters[i] = iters[iters.size() - 1];
+            iters.resize(iters.size() - 1);
+            --i;
+            continue;
         }
-        thread_pool_->AddTask(std::bind(&BlockManager::RemoveBlock, this, block_id));
+        int64_t tmp;
+        if (1 != sscanf(it->key().data(), "%ld", &tmp)) {
+            LOG(WARNING, "[FindSmallest] Unknown meta key: %s\n",
+                it->key().ToString().c_str());
+            return -1;
+        }
+        if (tmp < id) {
+            id = tmp;
+            *idx = i;
+        }
     }
-    delete it;
-    return true;
+    assert(id != LLONG_MAX);
+    return id;
 }
 
-}
-}
-
-/* vim: set expandtab ts=4 sw=4 sts=4 tw=100: */
+} // namespace bfs
+} // namespace baidu
