+++ src/chunkserver/block_manager.h
@@ -1,8 +1,7 @@
-// Copyright (c) 2015, Baidu.com, Inc. All Rights Reserved.
+// Copyright (c) 2016, Baidu.com, Inc. All Rights Reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
-// Author: yanshiguang02@baidu.com
 
 #ifndef  BAIDU_BFS_BLOCK_MANAGER_H_
 #define  BAIDU_BFS_BLOCK_MANAGER_H_
@@ -17,6 +16,7 @@
 
 namespace leveldb {
 class DB;
+class Iterator;
 }
 
 namespace baidu {
@@ -25,38 +25,34 @@ namespace bfs {
 class BlockMeta;
 class Block;
 class FileCache;
+class Disk;
 
 class BlockManager {
 public:
     BlockManager(const std::string& store_path);
     ~BlockManager();
     int64_t DiskQuota()  const;
-    /// Load meta from disk
     bool LoadStorage();
     int64_t NameSpaceVersion() const;
     bool SetNameSpaceVersion(int64_t version);
-    int64_t ListBlocks(std::vector<BlockMeta>* blocks, int64_t offset, int32_t num);
-    Block* CreateBlock(int64_t block_id, int64_t* sync_time, StatusCode* status);
-    Block* FindBlock(int64_t block_id);
+    int64_t ListBlocks(std::vector<BlockMeta>* blocks, int64_t offset, uint32_t num);
+    Block* CreateBlock(int64_t block_id,  StatusCode* status);
     bool CloseBlock(Block* block);
     StatusCode RemoveBlock(int64_t block_id);
-    bool RemoveAllBlocks();
+    bool CleanUp(int64_t namespace_version);
+
+    Block* FindBlock(int64_t block_id);
+    bool AddBlock(int64_t block_id, Block* block);
+
 private:
-    std::string BlockId2Str(int64_t block_id);
-    bool RemoveBlockMeta(int64_t block_id);
-    bool RemoveAllBlocksAsync();
     void CheckStorePath(const std::string& store_path);
-    const std::string& GetStorePath(int64_t block_id);
-    bool SyncBlockMeta(const BlockMeta& meta, int64_t* sync_time);
+    Disk* PickDisk(int64_t block_id);
+    int64_t FindSmallest(std::vector<leveldb::Iterator*>& iters, int32_t* idx);
 private:
-    ThreadPool* thread_pool_;
-    std::vector<std::string> store_path_list_;
-    typedef std::map<int64_t, Block*> BlockMap;
-    BlockMap  block_map_;
-    leveldb::DB* metadb_;
+    std::vector<Disk*> disks_;
     FileCache* file_cache_;
     Mutex   mu_;
-    int64_t namespace_version_;
+    std::map<int64_t, Block*> block_map_;
     int64_t disk_quota_;
 };
